Cleanup and consolidation
