+++ platformio.ini
@@ -179,14 +179,14 @@ build_flags =
 extends = esp32_base
 board = heltec_wifi_lora_32_V2
 build_flags = 
-  ${esp32_base.build_flags} -D HELTEC_V2_0 -I variants/heltec_v2_1
+  ${esp32_base.build_flags} -D HELTEC_V2_0 -I variants/heltec_v2
 
 [env:heltec-v2.1]
 ;build_type = debug ; to make it possible to step through our jtag debugger 
 extends = esp32_base
 board = heltec_wifi_lora_32_V2
 build_flags = 
-  ${esp32_base.build_flags} -D HELTEC_V2_1 -I variants/heltec_v2_1
+  ${esp32_base.build_flags} -D HELTEC_V2_1 -I variants/heltec_v2
 
 [env:tlora-v1]
 extends = esp32_base
