--- variants/tlora_v1/variant.cpp
@@ -1,6 +1,0 @@
-#include "variant.h"
-
-void initVariant()
-{
-	// Empty for right now but this could be used for bootstrapping logic 
-}
